OCT_flask_APP-master/requirements.txt
absl-py==2.1.0
accelerate==0.30.1
albumentations @ git+https://github.com/albumentations-team/albumentations@f7f9596f6adf5e53bc15555fb264a065919773c9
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
async-timeout==4.0.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.0
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1707444604970/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dm-tree==0.1.8
dnspython==2.6.1
efficientnet_pytorch==0.7.1
email_validator==2.1.1
exceptiongroup==1.2.0
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.13.1
Flask==3.0.2
Flask-Avatars==0.2.3
Flask-Caching==2.1.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
flask-paginate==2024.4.12
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.7
fonttools==4.49.0
fsspec==2024.3.0
gast==0.4.0
gevent==24.2.1
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h5py==3.10.0
hiredis==2.3.2
huggingface-hub==0.23.0
idna==3.6
imageio==2.34.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1709821103657/work
importlib_resources==6.3.0
imutils==0.5.4
ipykernel @ file:///D:/bld/ipykernel_1708996677248/work
ipython==8.18.1
itsdangerous==2.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
joblib==1.3.2
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257377578/work
keract==4.5.1
keras==2.13.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader==0.3
libclang==18.1.1
Mako==1.3.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
munch==4.0.0
namex==0.0.7
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas==1.1.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
plotly==5.22.0
pretrainedmodels==0.7.4
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1705722541723/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq @ file:///D:/bld/pyzmq_1701783320990/work
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.10.1
seaborn==0.10.0
segmentation-models-pytorch==0.3.3
shortuuid==1.0.13
six==1.15.0
SQLAlchemy==2.0.28
SQLAlchemy-serializer==1.4.12
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12
tenacity==8.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0rc1
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.2
torch==1.10.1+cu102
torchaudio==0.10.1+cu102
torchvision==0.11.2+cu102
tornado @ file:///D:/bld/tornado_1708363257374/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1710254411456/work
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.1
wrapt==1.12.1
WTForms==3.1.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
zope.event==5.0
zope.interface==6.2
